+++ query/CHANGELOG.md
@@ -17,6 +17,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+* Methods that return/map step results now include results from _only the last attempt_ where there have been retries ([#1631](https://github.com/cucumber/common/pull/1631)). Affects methods:
+  * `Query#getPickleStepAttachments`
+  * `Query#getPickleStepTestStepResults`
+  * `Query#getPickleTestStepResults`
+  * `Query#getStatusCounts`
+  * `Query#getTestStepResults`
+  * `Query#getTestStepsAttachments`
+
+## [10.1.0] - 2021-05-31
+
+### Added
+
+* New `Query#getStatusCounts(pickleIds: string[])` method which calculates a summary
+  of a run.
+
 ## [10.0.0] - 2021-05-17
 
 ### Changed
@@ -131,7 +146,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * First JavaScript implementation
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/cucumber/compare/query/v10.0.0...main
+[Unreleased]: https://github.com/cucumber/cucumber/compare/query/v10.1.0...main
+[10.1.0]:      https://github.com/cucumber/cucumber/compare/cucumber-query/v10.0.0...query/v10.1.0
 [10.0.0]:      https://github.com/cucumber/cucumber/compare/cucumber-query/v9.0.2...query/v10.0.0
 [9.0.2]:      https://github.com/cucumber/cucumber/compare/cucumber-query/v9.0.1...query/v9.0.2
 [9.0.1]:      https://github.com/cucumber/cucumber/compare/cucumber-query/v9.0.0...query/v9.0.1